# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SaaS Platform API - Multi-tenant platform with auth, billing, and app registry
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCurrentUser
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/me/switch-org:
    post:
      operationId: switchOrganization
      tags: [users]
      summary: Switch active organization context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchOrgRequest"
      responses:
        "200":
          description: Switched organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"

  # ── ORGANIZATIONS ─────────────────────────────────────────────────────────
  /organizations:
    get:
      operationId: getMyOrganizations
      tags: [organizations]
      summary: Get organizations the current user belongs to
      responses:
        "200":
          description: Organizations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create a new organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          description: Created organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{orgId}:
    get:
      operationId: getOrganization
      tags: [organizations]
      summary: Get organization details
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    patch:
      operationId: updateOrganization
      tags: [organizations]
      summary: Update organization settings
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationRequest"
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{orgId}/members:
    get:
      operationId: getOrgMembers
      tags: [memberships]
      summary: Get organization members
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrgMember"
  /organizations/{orgId}/members/{userId}:
    delete:
      operationId: removeOrgMember
      tags: [memberships]
      summary: Remove a member from organization
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    patch:
      operationId: updateOrgMemberRole
      tags: [memberships]
      summary: Update member role in organization
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRoleRequest"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgMember"

  # ── INVITATIONS ────────────────────────────────────────────────────────────
...
  /organizations/{orgId}/apps:
    get:
      operationId: getOrgApps
      tags: [apps]
      summary: Get apps subscribed to by an organization
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Org apps list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrgApp"

  # ── SUBSCRIPTIONS ─────────────────────────────────────────────────────────
...
  /admin/organizations:
    get:
      operationId: adminGetOrganizations
      tags: [admin]
      summary: Admin - list all organizations
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: All organizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgPage"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required: [error]

    # ── AUTH ──────────────────────────────────────────────────────────────────
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isSuperAdmin:
          type: boolean
        activeOrgId:
          type: ["string", "null"]
        activeOrg:
          $ref: "#/components/schemas/Organization"
        memberships:
          type: array
          items:
            $ref: "#/components/schemas/OrgMembershipSummary"
      required: [id, email, isSuperAdmin]
...
    OrgMembershipSummary:
      type: object
      properties:
        orgId:
          type: string
        orgName:
          type: string
        orgSlug:
          type: string
        role:
          type: string
      required: [orgId, orgName, orgSlug, role]

    # ── USERS ─────────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isSuperAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, email, isSuperAdmin, createdAt]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
    SwitchOrgRequest:
      type: object
      properties:
        orgId:
          type: string
      required: [orgId]
...
    Organization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        memberCount:
          type: integer
        stripeCustomerId:
          type: ["string", "null"]
      required: [id, name, slug, createdAt]
    CreateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        website:
          type: string
      required: [name, slug]
    UpdateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
        website:
          type: string
        logoUrl:
          type: string
    OrgPage:
      type: object
      properties:
        organizations:
          type: array
          items:
            $ref: "#/components/schemas/Organization"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required: [organizations, total, limit, offset]

    # ── MEMBERSHIPS ───────────────────────────────────────────────────────────
    OrgMember:
      type: object
      properties:
        userId:
          type: string
        orgId:
          type: string
        role:
          type: string
          enum: [owner, admin, member, viewer]
        email:
          type: string
        name:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        joinedAt:
          type: string
          format: date-time
      required: [userId, orgId, role, email, joinedAt]
    UpdateMemberRoleRequest:
      type: object
      properties:
        role:
          type: string
          enum: [owner, admin, member, viewer]
      required: [role]

    # ── INVITATIONS ────────────────────────────────────────────────────────────
...
    OrgApp:
      type: object
      properties:
        appId:
          type: string
        appName:
          type: string
        appSlug:
          type: string
        status:
          type: string
          enum: [active, trialing, past_due, canceled, inactive]
        subscriptionId:
          type: ["string", "null"]
      required: [appId, appName, appSlug, status]
...