# lib/api-zod/src/generated/types/orgApp.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaaS Platform API - Multi-tenant platform with auth, billing, and app registry
 * OpenAPI spec version: 0.1.0
 */
import type { OrgAppStatus } from "./orgAppStatus";

export interface OrgApp {
  appId: string;
  appName: string;
  appSlug: string;
  status: OrgAppStatus;
  /** @nullable */
  subscriptionId?: string | null;
}